use shared component for inviting users
